requirements.txt
backtrader==1.9.78.123
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
fonttools==4.57.0
frozendict==2.4.6
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
joblib==1.4.2
kiwisolver==1.4.8
matplotlib==3.10.1
multitasking==0.0.11
numpy==2.2.4
packaging==24.2
pandas==2.2.3
peewee==3.17.9
pillow==11.1.0
platformdirs==4.3.7
pyaml==25.1.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
scikit-learn==1.6.1
scikit-optimize==0.10.2
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
sklearn-compat==0.1.3
soupsieve==2.6
ta==0.11.0
threadpoolctl==3.6.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
yfinance==0.2.55
